Added SOAP 1.2 capability. Now Sandesha2 handles SOAP versions more accurately. In the server side, the SOAP version of the response messages depend on the versions of respective request messages. In the client side, the SOAP version should be set by the client. sender.setSoapVersionURI(SOAP12Constants.SOAP_ENVELOPE_NAMESPACE_URI); Had to change the RM elements to get the SOAPFactory as a parameter rather tham deciding it.
Changed code to take Policy values from a dynamic policy bean. (filling this bean as to be done by a policy manager which is not yet complete).

RMClient can set AckRequested part depending on a requirement (not decided yet). The idintifier for this will be set by createSequenceResponseProcessor.

Many bug fixes and improvements.