Changed the methods of the MessageProcessor interface to throw AxisFaults.
SandeshaoutHandler was changed not to pause messages with wsa:anonymous endpoint as the wsa:to address. 
This will allow at least a sync response to be done (but no retransmissions will be possible).

A bug fix - CreateSequence should have a different replyTo from the application message.

Some bug fixes in close sequence and terminate sequence processing logic.