Made the property copying logic (from referenceMessage to newMessage) generic. Properties to be copied can be specified in the module.xml within the "propertiesToCopyFromReferenceMessage" property. If it is in the key1:key2 format, the property with the key 'key1' will be copied with the key 'key2'.
Fixed a bug that was causing random test failuers (polling messages having server side set to true causing the sync responses
to be ignored in the SenderWorker).

Added the goal secure:rampart:create which causes some services to be built which are configured according to Rampart.

Added a way to ask Sandesha2 to avoid automatic termination of sequences (I.e. when all acks are received).