Fixed the empty body LastMessage processing.
In sync scenarios the server will respond this with a LastMessage for the response sequence. But this
should not be an response to the original message, I.e. there should not be an relatesTo.

A fix to throw an exception when offer it not given when it is required.

A couple of other big fixes.