A correction to set the messageType of empty body LastMessages correctly.
Added a new MessageType for 'Polling Response Messages'.

Made LastMessage Response and Polling Response messages OutOnly. So they will go with OutOnly operations. this should be the case 
since these does not relate to the request message.

Removed the duplicate MessageRetransmissionAdjuster.adjustRetransmittion call in the MakeConnectionProcessor. This get called in the
SenderWorker.

Stopped the FaultManager from throwing out Faults directly. This causes problems in scenarios such as Ack piggybacked Application msgs.


A fix to transfer RequestResponsTransport object from the request message to the PollingResponse and LastMessageResponse messages. 
SenderWorker was fixed to pick this correctly.